+++ tcod/path.c
@@ -47,19 +47,28 @@ float PathCostArrayUInt32(
   return (float)(*(uint32_t*)pick_array_pointer(map, x2, y2));
 }
 
-static bool array2d_in_range(const struct NArray4* arr, int i, int j)
+static bool array_in_range(const struct NArray* arr, int n, const int* index)
+{
+  for (int i = 0; i < n; ++i) {
+    if (index[i] < 0 || index[i] >= arr->shape[i]) { return 0; }
+  }
+  return 1;
+}
+static bool array2d_in_range(const struct NArray* arr, int i, int j)
 {
   return 0 <= i && i < arr->shape[0] && 0 <= j && j < arr->shape[1];
 }
-
-static void* get_array2d_at(const struct NArray4* arr, int i, int j)
+static void* get_array_ptr(const struct NArray* arr, int n, const int* index)
 {
-  return arr->data + arr->strides[0] * i + arr->strides[1] * j;
+  unsigned char* ptr = (unsigned char*)arr->data;
+  for (int i = 0; i < n; ++i) {
+    ptr += arr->strides[i] * index[i];
+  }
+  return (void*)ptr;
 }
-
-static int64_t get_array2d_int64(const struct NArray4* arr, int i, int j)
+static int64_t get_array_int64(const struct NArray* arr, int n, const int* index)
 {
-  const void* ptr = get_array2d_at(arr, i, j);
+  const void* ptr = get_array_ptr(arr, n, index);
   switch (arr->type) {
     case np_int8:
       return *(const int8_t*)ptr;
@@ -81,15 +90,14 @@ static int64_t get_array2d_int64(const struct NArray4* arr, int i, int j)
       return 0;
   }
 }
-static int get_array2d_int(const struct NArray4* arr, int i, int j)
+static int get_array_int(const struct NArray* arr, int n, const int* index)
 {
-  return (int)get_array2d_int64(arr, i, j);
+  return (int)get_array_int64(arr, n, index);
 }
-
-
-static void set_array2d_int64(struct NArray4* arr, int i, int j, int64_t value)
+static void set_array_int64(
+    struct NArray* arr, int n, const int* index, int64_t value)
 {
-  void* ptr = get_array2d_at(arr, i, j);
+  void* ptr = get_array_ptr(arr, n, index);
   switch (arr->type) {
     case np_int8:
       *(int8_t*)ptr = (int8_t)value;
@@ -118,13 +126,22 @@ static void set_array2d_int64(struct NArray4* arr, int i, int j, int64_t value)
       return;
   }
 }
-static void set_array2d_int(struct NArray4* arr, int i, int j, int value)
+static void set_array2d_int64(struct NArray* arr, int i, int j, int64_t value)
+{
+  int index[2] = {i, j};
+  set_array_int64(arr, 2, index, value);
+}
+static void set_array_int(struct NArray* arr, int n, const int* index, int value)
+{
+  set_array_int64(arr, n, index, value);
+}
+static void set_array2d_int(struct NArray* arr, int i, int j, int value)
 {
   set_array2d_int64(arr, i, j, value);
 }
-static int64_t get_array2d_is_max(const struct NArray4* arr, int i, int j)
+static int64_t get_array_is_max(const struct NArray* arr, int n, const int* index)
 {
-  const void* ptr = get_array2d_at(arr, i, j);
+  const void* ptr = get_array_ptr(arr, n, index);
   switch (arr->type) {
     case np_int8:
       return *(const int8_t*)ptr == SCHAR_MAX;
@@ -152,43 +169,42 @@ static const int DIAGONAL_[4][2] = {{-1, -1}, {1, -1}, {-1, 1}, {1, 1}};
 
 static void dijkstra2d_add_edge(
     struct TCOD_Frontier* frontier,
-    struct NArray4* dist_array,
-    const struct NArray4* cost,
+    struct NArray* dist_array,
+    const struct NArray* cost,
     int edge_cost,
     const int dir[2])
 {
   const int index[2] = {
       frontier->active_index[0] + dir[0], frontier->active_index[1] + dir[1]
   };
-  if (!array2d_in_range(dist_array, index[0], index[1])) { return; }
-  edge_cost *= get_array2d_int(cost, index[0], index[1]);
+  if (!array_in_range(dist_array, 2, index)) { return; }
+  edge_cost *= get_array_int(cost, 2, index);
   if (edge_cost <= 0) { return; }
   int distance = frontier->active_dist + edge_cost;
-  if (get_array2d_int(dist_array, index[0], index[1]) <= distance) { return; }
-  set_array2d_int(dist_array, index[0], index[1], distance);
+  if (get_array_int(dist_array, 2, index) <= distance) { return; }
+  set_array_int(dist_array, 2, index, distance);
   TCOD_frontier_push(frontier, index, distance, distance);
 }
 
 int dijkstra2d(
-    struct NArray4* dist_array,
-    const struct NArray4* cost,
+    struct NArray* dist_array,
+    const struct NArray* cost,
     int edges_2d_n,
-    int* edges_2d)
+    const int* edges_2d)
 {
   struct TCOD_Frontier* frontier = TCOD_frontier_new(2);
   if (!frontier) { return TCOD_E_ERROR; }
   for (int i = 0; i < dist_array->shape[0]; ++i) {
     for (int j = 0; j < dist_array->shape[1]; ++j) {
-      if (get_array2d_is_max(dist_array, i, j)) { continue; }
       const int index[2] = {i, j};
-      int dist = get_array2d_int(dist_array, i, j);
+      if (get_array_is_max(dist_array, 2, index)) { continue; }
+      int dist = get_array_int(dist_array, 2, index);
       TCOD_frontier_push(frontier, index, dist, dist);
     }
   }
   while (TCOD_frontier_size(frontier)) {
     TCOD_frontier_pop(frontier);
-    int distance_here = get_array2d_int(
-        dist_array, frontier->active_index[0], frontier->active_index[1]);
+    int distance_here = get_array_int(dist_array, 2, frontier->active_index);
     if (frontier->active_dist != distance_here) { continue; }
     for (int i = 0; i < edges_2d_n; ++i) {
       dijkstra2d_add_edge(
@@ -199,8 +215,8 @@ int dijkstra2d(
 }
 
 int dijkstra2d_basic(
-    struct NArray4* dist_array,
-    const struct NArray4* cost,
+    struct NArray* dist_array,
+    const struct NArray* cost,
     int cardinal,
     int diagonal)
 {
@@ -208,16 +224,15 @@ int dijkstra2d_basic(
   if (!frontier) { return TCOD_E_ERROR; }
   for (int i = 0; i < dist_array->shape[0]; ++i) {
     for (int j = 0; j < dist_array->shape[1]; ++j) {
-      if (get_array2d_is_max(dist_array, i, j)) { continue; }
       const int index[2] = {i, j};
-      int dist = get_array2d_int(dist_array, i, j);
+      if (get_array_is_max(dist_array, 2, index)) { continue; }
+      int dist = get_array_int(dist_array, 2, index);
       TCOD_frontier_push(frontier, index, dist, dist);
     }
   }
   while (TCOD_frontier_size(frontier)) {
     TCOD_frontier_pop(frontier);
-    int distance_here = get_array2d_int(
-        dist_array, frontier->active_index[0], frontier->active_index[1]);
+    int distance_here = get_array_int(dist_array, 2, frontier->active_index);
     if (frontier->active_dist != distance_here) { continue; }
     if (cardinal > 0) {
       dijkstra2d_add_edge(frontier, dist_array, cost, cardinal, CARDINAL_[0]);
@@ -235,15 +250,15 @@ int dijkstra2d_basic(
   return TCOD_E_OK;
 }
 static void hillclimb2d_check_edge(
-    const struct NArray4* dist_array,
+    const struct NArray* dist_array,
     int* distance_in_out,
     const int origin[2],
     const int dir[2],
     int index_out[2])
 {
   const int next[2] = {origin[0] + dir[0], origin[1] + dir[1]};
-  if (!array2d_in_range(dist_array, next[0], next[1])) { return; }
-  const int next_distance = get_array2d_int(dist_array, next[0], next[1]);
+  if (!array_in_range(dist_array, 2, next)) { return; }
+  const int next_distance = get_array_int(dist_array, 2, next);
   if (next_distance < *distance_in_out) {
     *distance_in_out = next_distance;
     index_out[0] = next[0];
@@ -251,16 +266,16 @@ static void hillclimb2d_check_edge(
   }
 }
 int hillclimb2d(
-    const struct NArray4* dist_array,
+    const struct NArray* dist_array,
     int start_i,
     int start_j,
     int edges_2d_n,
-    int* edges_2d,
+    const int* edges_2d,
     int* out)
 {
-  int old_dist = get_array2d_int(dist_array, start_i, start_j);
-  int new_dist = old_dist;
   int next[2] = {start_i, start_j};
+  int old_dist = get_array_int(dist_array, 2, next);
+  int new_dist = old_dist;
   int length = 0;
   while (1) {
     ++length;
@@ -281,16 +296,16 @@ int hillclimb2d(
   }
 }
 int hillclimb2d_basic(
-    const struct NArray4* dist_array,
+    const struct NArray* dist_array,
     int start_i,
     int start_j,
     bool cardinal,
     bool diagonal,
     int* out)
 {
-  int old_dist = get_array2d_int(dist_array, start_i, start_j);
-  int new_dist = old_dist;
   int next[2] = {start_i, start_j};
+  int old_dist = get_array_int(dist_array, 2, next);
+  int new_dist = old_dist;
   int length = 0;
   while (1) {
     ++length;
@@ -317,4 +332,128 @@ int hillclimb2d_basic(
     }
     old_dist = new_dist;
   }
+}
+void path_compute_add_edge(
+    struct TCOD_Frontier* frontier,
+    struct NArray* dist_map,
+    struct NArray* travel_map,
+    const struct NArray* cost_map,
+    const int* edge_rule)
+{
+  int dest[TCOD_PATHFINDER_MAX_DIMENSIONS];
+  for (int i = 0; i < frontier->ndim; ++i) {
+    dest[i] = frontier->active_index[i] + edge_rule[i];
+  }
+  if (!array_in_range(dist_map, frontier->ndim, dest)) { return; }
+  int edge_cost = edge_rule[frontier->ndim];
+  edge_cost *= get_array_int(cost_map, frontier->ndim, dest);
+  if (edge_cost <= 0) { return; }
+  int distance = frontier->active_dist + edge_cost;
+  if (get_array_int(dist_map, frontier->ndim, dest) <= distance) { return; }
+  set_array_int(dist_map, frontier->ndim, dest, distance);
+  int* path = get_array_ptr(travel_map, frontier->ndim, dest);
+  for (int i = 0; i < frontier->ndim; ++i) {
+    path[i] = frontier->active_index[i];
+  }
+  TCOD_frontier_push(frontier, dest, distance, distance);
+}
+
+int path_compute_step(
+    struct TCOD_Frontier* frontier,
+    struct NArray* dist_map,
+    struct NArray* travel_map,
+    int n,
+    const struct PathfinderRule* rules)
+{
+  if (!frontier) {
+    return TCOD_set_errorv("Missing frontier.");
+  }
+  if (frontier->ndim <= 0 || frontier->ndim > TCOD_PATHFINDER_MAX_DIMENSIONS) {
+    return TCOD_set_errorv("Invalid frontier->ndim.");
+  }
+  if (!dist_map) {
+    return TCOD_set_errorv("Missing dist_map.");
+  }
+  if (frontier->ndim != dist_map->ndim) {
+    return TCOD_set_errorv("Invalid or corrupt input.");
+  }
+  if (travel_map && frontier->ndim + 1 != travel_map->ndim) {
+    return TCOD_set_errorv("Invalid or corrupt input.");
+  }
+  TCOD_frontier_pop(frontier);
+  for (int i = 0; i < n; ++i) {
+    if (rules[i].condition.type) {
+      if (!get_array_int(&rules[i].condition, frontier->ndim, frontier->active_index)) {
+        continue;
+      }
+    }
+    for (int edge_i = 0; edge_i < rules[i].edge_count; ++edge_i) {
+      path_compute_add_edge(
+          frontier,
+          dist_map,
+          travel_map,
+          &rules[i].cost,
+          &rules[i].edge_array[edge_i * (frontier->ndim + 1)]);
+    }
+  }
+  return 0;
+}
+int path_compute(
+    struct TCOD_Frontier* frontier,
+    struct NArray* dist_map,
+    struct NArray* travel_map,
+    int n,
+    const struct PathfinderRule* rules)
+{
+  if (!frontier) {
+    return TCOD_set_errorv("Missing frontier.");
+  }
+  while (TCOD_frontier_size(frontier)) {
+    int err = path_compute_step(
+        frontier,
+        dist_map,
+        travel_map,
+        n,
+        rules);
+    if (err < 0) { return err; }
+  }
+  return 0;
+}
+size_t get_travel_path(
+    int8_t ndim, const struct NArray* travel_map, const int* start, int* out)
+{
+  if (ndim <= 0 || ndim > TCOD_PATHFINDER_MAX_DIMENSIONS) {
+    return TCOD_set_errorv("Invalid ndim.");
+  }
+  if (!travel_map) {
+    return TCOD_set_errorv("Missing travel_map.");
+  }
+  if (!start) {
+    return TCOD_set_errorv("Missing start.");
+  }
+  if (ndim != travel_map->ndim - 1) {
+    return TCOD_set_errorv("Invalid or corrupt input.");
+  }
+  const int* next = get_array_ptr(travel_map, ndim, start);
+  const int* current = start;
+  size_t max_loops = 1;
+  size_t length = 0;
+  for (int i = 0; i < ndim; ++i) { max_loops *= (size_t)travel_map->shape[i]; }
+  while (current != next) {
+    ++length;
+    if (out) {
+      for (int i = 0; i < ndim; ++i) { out[i] = current[i]; }
+      out += ndim;
+    }
+    current = next;
+    if (!array_in_range(travel_map, ndim, next)) {
+      return TCOD_set_errorvf(
+          "Index (%i, %i) is out of range.", next[0], next[1]);
+    }
+    next = get_array_ptr(travel_map, ndim, next);
+    if (!out && length == max_loops) {
+      return TCOD_set_errorv("Possible cyclic loop detected.");
+    }
+  }
+  return length;
 }